Override apache default modules and adjust vhost.
I wanted indexes simply removed from all the servers. However there is
no easy way to remove an enabled module. The best we can do is prevent
it from installing initially. Further, I don't want the default vhosts
file to come in with a prepended 15 since we are defaulting this to be
named `000-default` anyway to get it to the top of the priority list.