Allow the 'expose_php' setting to be configured.
A nice security hardening measure is setting 'expose_php' to
'Off' in the apache php.ini file. This prevents the returned
headers from revealing information about PHP and the server
version information. Previously it was not possible to set this
through the drupal-php module since the option in the required
php module required further configuration.